refactor and upgrade simple_3d example
